major changes. must be merged to devel.

+++ sandy/smpling.py
@@ -14,7 +14,6 @@ import sys
 import os
 import multiprocessing as mp
 from copy import deepcopy
-import re
 import shutil
 
 
@@ -45,44 +44,49 @@ def sample_xs(tape, NSMP):
     DfPert.columns.name = 'SMP'
     return DfPert
 
-
-def perturb(XsSeries, PertSeries):
-    Pert = e6.pandas_interpolate(PertSeries,
-                           np.unique(XsSeries.index).tolist(),
-                           axis="rows")
-    Xs = XsSeries.multiply(Pert, axis="index")
-    Xs.name = XsSeries.name # this is mt
-    # Negative values are set to zero
-    Xs[Xs <= 0] = 0
-    return Xs
-
-
 def perturb_xs(tape, PertSeriesXs):
     Xs = e6.extract_xs(tape)
     for mat, mt in Xs:
         if mat not in PertSeriesXs.index.get_level_values("MAT").unique():
             continue
         mtListPert = PertSeriesXs.loc[mat].index.get_level_values("MT").unique()
-#        mtListXs = Xs[mat].columns
         if mt in mtListPert:
             mtPert = mt
-        elif mt in range(800,850) and 107 in mtListPert:
+        elif mt in range(800,850) \
+        and not list(filter(lambda x: x in mtListPert, range(800,850))) \
+        and 107 in mtListPert:
             mtPert = 107
-        elif mt in range(750,800) and 106 in mtListPert:
+        elif mt in range(750,800) \
+        and not list(filter(lambda x: x in mtListPert, range(750,800))) \
+        and 106 in mtListPert:
             mtPert = 106
-        elif mt in range(700,750) and 105 in mtListPert:
+        elif mt in range(700,750) \
+        and not list(filter(lambda x: x in mtListPert, range(700,750))) \
+        and 105 in mtListPert:
             mtPert = 105
-        elif mt in range(650,700) and 104 in mtListPert:
+        elif mt in range(650,700) \
+        and not list(filter(lambda x: x in mtListPert, range(650,700))) \
+        and 104 in mtListPert:
             mtPert = 104
-        elif mt in range(600,650) and 103 in mtListPert:
+        elif mt in range(600,650) \
+        and not list(filter(lambda x: x in mtListPert, range(600,650))) \
+        and 103 in mtListPert:
             mtPert = 103
-        elif mt in range(102,118) and 101 in mtListPert:
+        elif mt in range(102,118) \
+        and not list(filter(lambda x: x in mtListPert, range(102,118))) \
+        and 101 in mtListPert:
             mtPert = 101
-        elif mt in (19,20,21,38) and 18 in mtListPert:
+        elif mt in (19,20,21,38) \
+        and not list(filter(lambda x: x in mtListPert, (19,20,21,38))) \
+        and 18 in mtListPert:
             mtPert = 18
-        elif mt in (18,101) and 27 in mtListPert:
+        elif mt in (18,101) \
+        and not list(filter(lambda x: x in mtListPert, (18,101))) \
+        and 27 in mtListPert:
             mtPert = 27
-        elif mt in range(50,92) and 4 in mtListPert:
+        elif mt in range(50,92) \
+        and not list(filter(lambda x: x in mtListPert, range(50,92))) \
+        and 4 in mtListPert:
             mtPert = 4
         else:
             continue
@@ -91,45 +95,8 @@ def perturb_xs(tape, PertSeriesXs):
         Xs[mat,mt] = Xs[mat,mt].multiply(P, axis="index")
         # Negative values are set to zero
         Xs[mat,mt][Xs[mat,mt] <= 0] = 0
-    for mat in Xs.columns.get_level_values("MAT").unique():
-        # Redundant XS
-        daughters = [ x for x in range(800,850) if x in Xs[mat].columns]
-        if daughters:
-            Xs[mat][107] = Xs[mat][daughters].sum(axis=1)
-        daughters = [ x for x in range(750,800) if x in Xs[mat].columns]
-        if daughters:
-            Xs[mat][106] = Xs[mat][daughters].sum(axis=1)
-        daughters = [ x for x in range(700,750) if x in Xs[mat].columns]
-        if daughters:
-            Xs[mat][105] = Xs[mat][daughters].sum(axis=1)
-        daughters = [ x for x in range(650,700) if x in Xs[mat].columns]
-        if daughters:
-            Xs[mat][104] = Xs[mat][daughters].sum(axis=1)
-        daughters = [ x for x in range(600,650) if x in Xs[mat].columns]
-        if daughters:
-            Xs[mat][103] = Xs[mat][daughters].sum(axis=1)
-        daughters = [ x for x in range(102,118) if x in Xs[mat].columns]
-        if daughters:
-            Xs[mat][101] = Xs[mat][daughters].sum(axis=1)
-        daughters = [ x for x in (19,20,21,38) if x in Xs[mat].columns]
-        if daughters:
-            Xs[mat][18] = Xs[mat][daughters].sum(axis=1)
-        daughters = [ x for x in (18,101) if x in Xs[mat].columns]
-        if daughters:
-            Xs[mat][27] = Xs[mat][daughters].sum(axis=1)
-        daughters = [ x for x in range(50,92) if x in Xs[mat].columns]
-        if daughters:
-            Xs[mat][4] = Xs[mat][daughters].sum(axis=1)
-        daughters = [ x for x in (4,5,11,*range(16,19),*range(22,38),41,42,44,45) if x in Xs[mat].columns]
-        if daughters:
-            Xs[mat][3] = Xs[mat][daughters].sum(axis=1)
-        daughters = [ x for x in (2,3) if x in Xs[mat].columns]
-        if daughters:
-            Xs[mat][1] = Xs[mat][daughters].sum(axis=1)
-        Xs[mat] = Xs[mat][mtListXs] # keep only mt present in the original file
-    tape = e6.update_xs(tape, Xs)
-    tape = e6.write_mf3_mt(tape)
-    e6.write_tape(tape, output)
+    Xs = e6.reconstruct_xs(Xs)
+    return e6.write_mf1_nubar( e6.write_mf3_mt( e6.update_xs(tape, Xs) ) )
 
 def perturb_chi(tape, PertSeriesChi):
     PertSeriesChi.name = 'SMP'
@@ -171,51 +138,61 @@ def sampling(tape, output, PertSeriesXs=None, PertSeriesChi=None):
         itape = perturb_chi(itape, PertSeriesChi)
     if PertSeriesXs is not None:
         itape = perturb_xs(itape, PertSeriesXs)
+
     e6.write_tape(itape, output)
 
 if __name__ == '__main__':
     __spec__ = None
-    #file = "data_test/H1.txt"
-    #file = "96-Cm-242g.jeff33"
-    #file = "nubar_endfb80.tape"
-    #file = "nubar_jeff33.tape"
-    #NSMP = 100
-    #file = "JEFF33-rdd_all.asc"
-    #file = "26-Fe-56g.jeff33"
     if len(sys.argv) == 1:
-        sys.argv.extend(["data_test\H1.txt",
-                         "--outdir", os.path.join("..","ttt"),
-                         "--njoy", r"J:\NEA\NDaST\NJOY\njoy2012_50.exe"])
+        sys.argv.extend(["data_test\96-Cm-242g.jeff33",
+                         "--outdir", os.path.join("..","tmp-cm"),
+                         "--njoy", r"J:\NEA\NDaST\NJOY\njoy2012_50.exe",
+                         "-mf", "33", "-mf", "35"])
     settings.init()
     tape = e6.endf2df(settings.args.endf6)
+    if settings.args.keep_cov_mf:
+        query = "|".join([ "MF=={}".format(x) for x in settings.args.keep_cov_mf])
+        tape = tape.query("MF < 31 | ({})".format(query))
+    if settings.args.keep_cov_mt:
+        query = "|".join([ "MF=={}".format(x) for x in settings.args.keep_cov_mf])
+        tape = tape.query("MF < 31 | ({})".format(query))
+    MATS = list(tape.index.get_level_values("MAT").unique())
+    MFS = list(tape.index.get_level_values("MF").unique())
 
-    from sandy.njy import FileNJOY
-    fnjoy = FileNJOY()
-    # Write NJOY file
-    fnjoy.copy_to_tape(settings.args.endf6, 20, dst=settings.args.outdir)
-    fnjoy.reconr(20, 21, mat=tape.index.get_level_values('MAT').unique())
-    fnjoy.stop()
-    fnjoy.run(settings.args.njoy, cwd=settings.args.outdir)
-    settings.args.pendf = os.path.join(settings.args.outdir,
-                                       '{}.pendf'.format(os.path.basename(settings.args.endf6)))
-    shutil.move(os.path.join(settings.args.outdir, r'tape21'), settings.args.pendf)
-    # Remove NJOY junk outputs
-    os.unlink(os.path.join(settings.args.outdir, r'tape20'))
-    os.unlink(os.path.join(settings.args.outdir, r'output'))
-    ptape = e6.endf2df(settings.args.pendf)
-
-
-    PertChi = sample_chi(tape, settings.args.samples)
     PertXs = sample_xs(tape, settings.args.samples)
+    PertChi = sample_chi(tape, settings.args.samples)
 
-    # Problem when running python from windows to linux
-#    outname = os.path.join(settings.args.outdir, os.path.basename(settings.args.endf6) + '-{}')
-#    sampling(tape, outname.format(1), PertSeriesChi=PertChi[1])
-
-    # Problem when running python from windows to linux
-    outname = os.path.join(settings.args.outdir, os.path.basename(settings.args.endf6) + '-{}')
-    sampling(ptape, outname.format(1), PertSeriesXs=PertXs[1])
-    #df_nu = extract_nu(tape) # dictionary (keys are MAT) of dataframes
+#    sampling(tape, 'temp.endf', PertSeriesXs=PertXs.query("MT==452 | MT==455 | MT==456")[1])
+#    sampling(ptape, 'temp.pendf', PertSeriesXs=PertXs.query("MT!=452 & MT!=455 & MT!=456")[1])
 
     pool = mp.Pool(processes=settings.args.processes)
-    [pool.apply( sampling, args=(tape, PertXs[ismp], outname.format(ismp))) for ismp in range(1,settings.args.samples+1) ]
+    outname = os.path.join(settings.args.outdir, os.path.basename(settings.args.endf6) + '-{}')
+    [ pool.apply(sampling,
+                 args = (tape, outname.format(ismp)),
+                 kwds = {"PertSeriesXs"  : PertXs[ismp] if not PertXs.empty() else None,
+                         "PertSeriesChi" : PertChi[ismp] if not PertChi.empty() else None}
+                 ) for ismp in range(1,settings.args.samples+1) ]
+
+    if 33 in MFS:
+        if not settings.args.pendf:
+            from sandy.njy import FileNJOY
+            fnjoy = FileNJOY()
+            # Write NJOY file
+            fnjoy.copy_to_tape(settings.args.endf6, 20, dst=settings.args.outdir)
+            fnjoy.reconr(20, 21, mat=tape.index.get_level_values('MAT').unique())
+            fnjoy.stop()
+            fnjoy.run(settings.args.njoy, cwd=settings.args.outdir)
+            settings.args.pendf = os.path.join(settings.args.outdir,
+                                               '{}.pendf'.format(os.path.basename(settings.args.endf6)))
+            shutil.move(os.path.join(settings.args.outdir, r'tape21'), settings.args.pendf)
+            # Remove NJOY junk outputs
+            os.unlink(os.path.join(settings.args.outdir, r'tape20'))
+            os.unlink(os.path.join(settings.args.outdir, r'output'))
+        ptape = e6.endf2df(settings.args.pendf)
+
+        pool = mp.Pool(processes=settings.args.processes)
+        outname = os.path.join(settings.args.outdir, os.path.basename(settings.args.pendf) + '-{}')
+        [ pool.apply(sampling,
+                     args = (ptape, outname.format(ismp)),
+                     kwds = {"PertSeriesXs" : PertXs[ismp] if not PertXs.empty() else None}
+                     ) for ismp in range(1,settings.args.samples+1) ]
